GW-22: Added about.js and refactored some codes
