+++ src/components/about/About.js
@@ -1,0 +1,21 @@
+import { Typography } from '@mui/material'
+import React from 'react'
+
+const About = () => {
+    return (
+        <>
+            <Typography variant='h3'>
+                WHO WE ARE
+            </Typography>
+            <Typography variant='body1'>
+                Hello! We are a non-profit cultural organization centered in promoting Indonesian hospitality around the UBC Vancouver campus.
+            </Typography>
+            <Typography variant='body1'>
+                GISAU aims to foster an inclusive, close-knitted, and connected community that exemplifies the signature Indonesian warmth and welcomes the diverse UBC society of Indonesian and non-Indonesian students alike.
+            </Typography>
+            <Button></Button>
+        </>
+    )
+}
+
+export default About
